Extract cio.track() specific functionality to an abstract class
